# clean_website/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Study Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /materials:
    get:
      operationId: listMaterials
      tags: [materials]
      summary: List materials, optionally filtered by courseId
      parameters:
        - name: courseId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Material"
    post:
      operationId: createMaterial
      tags: [materials]
      summary: Create/upload a material (text or URL)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
  /materials/{id}:
    get:
      operationId: getMaterial
      tags: [materials]
      summary: Get a material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
        "404":
          description: Not found
    delete:
      operationId: deleteMaterial
      tags: [materials]
      summary: Delete a material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Summaries ─────────────────────────────────────────────────────────────
  /materials/{id}/summaries:
    get:
      operationId: listSummaries
      tags: [summaries]
      summary: List summaries for a material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Summary"
    post:
      operationId: generateSummary
      tags: [summaries]
      summary: Generate a summary for a material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SummaryRequest"
      responses:
        "201":
          description: Generated summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Summary"
...
  /materials/{id}/flashcard-decks:
    get:
      operationId: listFlashcardDecks
      tags: [flashcards]
      summary: List flashcard decks for a material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Decks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FlashcardDeck"
    post:
      operationId: generateFlashcards
      tags: [flashcards]
      summary: Generate flashcard deck from material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlashcardRequest"
      responses:
        "201":
          description: Generated deck
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashcardDeck"
  /flashcard-decks/{id}:
    get:
      operationId: getFlashcardDeck
      tags: [flashcards]
      summary: Get a flashcard deck with cards
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deck with cards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashcardDeck"
    delete:
      operationId: deleteFlashcardDeck
      tags: [flashcards]
      summary: Delete a flashcard deck
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /flashcards/{id}/review:
    post:
      operationId: reviewFlashcard
      tags: [flashcards]
      summary: Record flashcard review result (spaced repetition)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlashcardReview"
      responses:
        "200":
          description: Updated card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flashcard"

  # ── Questions & Answers ───────────────────────────────────────────────────
...
components:
  schemas:
...
    Material:
      type: object
      required: [id, title, contentType, status, createdAt]
      properties:
        id:
          type: integer
        courseId:
          type: ["integer", "null"]
        title:
          type: string
        contentType:
          type: string
          enum: [text, url, youtube, pdf, docx, pptx, image, audio, video]
        status:
          type: string
          enum: [pending, processing, ready, error]
        language:
          type: string
          enum: [he, en, mixed]
        extractedText:
          type: ["string", "null"]
        sourceUrl:
          type: ["string", "null"]
        fileSize:
          type: ["integer", "null"]
        duration:
          type: ["integer", "null"]
        summaryCount:
          type: integer
        flashcardCount:
          type: integer
        questionCount:
          type: integer
        examCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    MaterialInput:
      type: object
      required: [title, contentType]
      properties:
        title:
          type: string
          minLength: 1
        courseId:
          type: integer
        contentType:
          type: string
          enum: [text, url, youtube, pdf, docx, pptx, image, audio, video]
        language:
          type: string
          enum: [he, en, mixed]
        text:
          type: string
        sourceUrl:
          type: string

    # ── Summary ──────────────────────────────────────────────────────────────
    Summary:
      type: object
      required: [id, materialId, summaryType, language, content, createdAt]
      properties:
        id:
          type: integer
        materialId:
          type: integer
        summaryType:
          type: string
          enum: [quick, detailed, chapter, topic, key_takeaways, exam_focused]
        language:
          type: string
          enum: [he, en]
        content:
          type: string
        keyPoints:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
    SummaryRequest:
      type: object
      required: [summaryType, language]
      properties:
        summaryType:
          type: string
          enum: [quick, detailed, chapter, topic, key_takeaways, exam_focused]
        language:
          type: string
          enum: [he, en]
        topic:
          type: string

    # ── Flashcards ────────────────────────────────────────────────────────────
    FlashcardDeck:
      type: object
      required: [id, materialId, title, language, createdAt]
      properties:
        id:
          type: integer
        materialId:
          type: integer
        title:
          type: string
        language:
          type: string
          enum: [he, en]
        cardCount:
          type: integer
        masteredCount:
          type: integer
        cards:
          type: array
          items:
            $ref: "#/components/schemas/Flashcard"
        createdAt:
          type: string
          format: date-time
    Flashcard:
      type: object
      required: [id, deckId, front, back, difficulty, createdAt]
      properties:
        id:
          type: integer
        deckId:
          type: integer
        front:
          type: string
        back:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        cardType:
          type: string
          enum: [qa, definition, formula, concept]
        reviewCount:
          type: integer
        nextReviewAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    FlashcardRequest:
      type: object
      required: [language]
      properties:
        language:
          type: string
          enum: [he, en]
        cardCount:
          type: integer
        cardTypes:
          type: array
          items:
            type: string
            enum: [qa, definition, formula, concept]
    FlashcardReview:
      type: object
      required: [result]
      properties:
        result:
          type: string
          enum: [again, hard, good, easy]

    # ── Questions ─────────────────────────────────────────────────────────────
...